Add a header area
The header area has a link to the projects Github, a save button, and a
load button. The buttons are colored and placed correctly. This creates
a clearner looking UI and better UX as the the save/load button is not
sitting at the bottom of the page.
Also, the title was made a bit smaller and the margin-top for it was
also decrease to 16px.